Rearrange order of checks according to best order of fixing issues
Check for trailing whitespace only works on files with Unix line endings. Line endings should be fixed before running the trailing whitespace check. When using the checks to fix formatting issues, it's logical to do the fixes in the order of the checks, so the checks should be in the optimal order of fixing issues.